Safari-609.1.11.1/LayoutTests/fast/media/mq-any-hover-invalid-expected.txt

Test invalid input with the any-hover media feature.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.matchMedia("(any-hover:)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-hover: )").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-hover: WebKit)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-hover: fine)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-hover: portrait)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-hover: \"None\")").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-hover: 'None')").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS successfullyParsed is true

TEST COMPLETE
